VisitorListener - Provide just the interface to capture which schemas are validating
